+++ p2panda-stream/src/stream/ingest.rs
@@ -153,11 +153,11 @@ where
             // 2. Validate and check the log-integrity of the incoming operation. If it is valid it
             //    get's persisted and the log optionally pruned.
             let ingest_fut = async {
-                let log_id: L = header
-                    .extract()
+                let log_id = header
+                    .extension()
                     .ok_or(IngestError::MissingHeaderExtension("log_id".into()))?;
                 let prune_flag: PruneFlag = header
-                    .extract()
+                    .extension()
                     .ok_or(IngestError::MissingHeaderExtension("prune_flag".into()))?;
                 ingest_operation::<S, L, E>(
                     &mut store,